chore: move constant to a dedicated file so we can share it with renderer part properly avoid to have renderer part depending on internals from configuration-registry related to https://github.com/containers/podman-desktop/pull/1605
