+++ packages/renderer/src/PreferencesNavigation.svelte
@@ -3,7 +3,7 @@ import { onMount } from 'svelte';
 import { Buffer } from 'buffer';
 import { extensionInfos } from './stores/extensions';
 import { configurationProperties } from './stores/configurationProperties';
-import { CONFIGURATION_DEFAULT_SCOPE } from '../../main/src/plugin/configuration-registry';
+import { CONFIGURATION_DEFAULT_SCOPE } from '../../main/src/plugin/configuration-registry-constants';
 import type { ProviderInfo } from '../../main/src/plugin/api/provider-info';
 import { providerInfos } from './stores/providers';
