+++ packages/renderer/src/lib/preferences/PreferencesRendering.svelte
@@ -1,8 +1,7 @@
 <script lang="ts">
-import {
-  CONFIGURATION_DEFAULT_SCOPE,
-  IConfigurationPropertyRecordedSchema,
-} from '../../../../main/src/plugin/configuration-registry';
+import type { IConfigurationPropertyRecordedSchema } from '../../../../main/src/plugin/configuration-registry';
+import { CONFIGURATION_DEFAULT_SCOPE } from '../../../../main/src/plugin/configuration-registry-constants';
+
 import PreferencesRenderingItem from './PreferencesRenderingItem.svelte';
 
 export let properties: IConfigurationPropertyRecordedSchema[] = [];
